adds optional conversion
don't like the control flow with the KeyError, but want to preserve the
original intent of using Hash#fetch for strict typing in postgres to
mysql type conversion.

changing in prep for allowing optional conversions on a per-column basis
to support custom data types in postgres (mint has a custom
mint_squareup_status enum).